Made various changes: - Made fbx files 10 times bigger. - Moved scene scale to meters. - Made some updates to the soldier prefab.
Commented out some stuff in the TurnManager script as the GameOver canvas has seemed to have been removed from the Main game scene. Also the Teddy scripts have been disabled as they were causing issues in the main scene. Not sure on some of the stuff added so I just commited everything, if there is any issues with the meshes then it might be for that reason.